daemon: extract owner-event policy machinery into events.rs
Pure move — DaemonEvent, its delivery policies, DaemonTelemetrySlot, the
OwnerEvent impl, DaemonEventSender, and the emit helpers relocate from
mod.rs (1332 lines, 8 under its size-ratchet cap) into a new events.rs so
the upcoming GUI-wiring streams can add their event lanes without busting
the pin.
